requirements.txt

fastapi
uvicorn[standard]
boto3
python-jose
PyYAML
requests
